# perf!: Prevent use of slow `str.contains_token`