--- a/examples/ndjson/parser.rs
+++ b/examples/ndjson/parser.rs
@@ -108,7 +108,7 @@ fn string<'i, E: ParseError<Stream<'i>> + ContextError<Stream<'i>, &'static str>
 /// You can mix the above declarative parsing with an imperative style to handle more unique cases,
 /// like escaping
 fn character<'i, E: ParseError<Stream<'i>>>(input: Stream<'i>) -> IResult<Stream<'i>, char, E> {
-    let (input, c) = none_of("\"").parse_next(input)?;
+    let (input, c) = none_of('"').parse_next(input)?;
     if c == '\\' {
         alt((
             any.verify_map(|c| {
@@ -201,7 +201,7 @@ fn ws<'i, E: ParseError<Stream<'i>>>(input: Stream<'i>) -> IResult<Stream<'i>, &
     take_while(0.., WS).parse_next(input)
 }
 
-const WS: &str = " \t";
+const WS: &[char] = &[' ', '\t'];
 
 #[cfg(test)]
 mod test {