--- a/examples/string/parser.rs
+++ b/examples/string/parser.rs
@@ -80,7 +80,7 @@ where
 fn parse_literal<'a, E: ParseError<&'a str>>(input: &'a str) -> IResult<&'a str, &'a str, E> {
     // `take_till1` parses a string of 0 or more characters that aren't one of the
     // given characters.
-    let not_quote_slash = take_till1("\"\\");
+    let not_quote_slash = take_till1(['"', '\\']);
 
     // `verify` runs a parser, then runs a verification function on the output of
     // the parser. The verification function accepts the output only if it