--- a/src/ascii/tests.rs
+++ b/src/ascii/tests.rs
@@ -571,7 +571,7 @@ mod complete {
         fn escaped_string(input: &str) -> IResult<&str, &str> {
             use crate::ascii::alpha0;
             use crate::token::one_of;
-            escaped(alpha0, '\\', one_of("n")).parse_next(input)
+            escaped(alpha0, '\\', one_of(['n'])).parse_next(input)
         }
 
         escaped_string("7").unwrap();
@@ -588,7 +588,11 @@ mod complete {
 
             delimited(
                 '"',
-                escaped(opt(none_of(r#"\""#)), '\\', one_of(r#"\"rnt"#)),
+                escaped(
+                    opt(none_of(['\\', '"'])),
+                    '\\',
+                    one_of(['\\', '"', 'r', 'n', 't']),
+                ),
                 '"',
             )
             .parse_next(input)
@@ -605,7 +609,7 @@ mod complete {
         use crate::token::one_of;
 
         fn esc(i: &[u8]) -> IResult<&[u8], &[u8]> {
-            escaped(alpha, '\\', one_of("\"n\\")).parse_next(i)
+            escaped(alpha, '\\', one_of(['\"', 'n', '\\'])).parse_next(i)
         }
         assert_eq!(esc(&b"abcd;"[..]), Ok((&b";"[..], &b"abcd"[..])));
         assert_eq!(esc(&b"ab\\\"cd;"[..]), Ok((&b";"[..], &b"ab\\\"cd"[..])));
@@ -629,7 +633,7 @@ mod complete {
         );
 
         fn esc2(i: &[u8]) -> IResult<&[u8], &[u8]> {
-            escaped(digit, '\\', one_of("\"n\\")).parse_next(i)
+            escaped(digit, '\\', one_of(['\"', 'n', '\\'])).parse_next(i)
         }
         assert_eq!(esc2(&b"12\\nnn34"[..]), Ok((&b"nn34"[..], &b"12\\n"[..])));
     }
@@ -641,7 +645,7 @@ mod complete {
         use crate::token::one_of;
 
         fn esc(i: &str) -> IResult<&str, &str> {
-            escaped(alpha, '\\', one_of("\"n\\")).parse_next(i)
+            escaped(alpha, '\\', one_of(['\"', 'n', '\\'])).parse_next(i)
         }
         assert_eq!(esc("abcd;"), Ok((";", "abcd")));
         assert_eq!(esc("ab\\\"cd;"), Ok((";", "ab\\\"cd")));
@@ -662,12 +666,12 @@ mod complete {
         );
 
         fn esc2(i: &str) -> IResult<&str, &str> {
-            escaped(digit, '\\', one_of("\"n\\")).parse_next(i)
+            escaped(digit, '\\', one_of(['\"', 'n', '\\'])).parse_next(i)
         }
         assert_eq!(esc2("12\\nnn34"), Ok(("nn34", "12\\n")));
 
         fn esc3(i: &str) -> IResult<&str, &str> {
-            escaped(alpha, '\u{241b}', one_of("\"n")).parse_next(i)
+            escaped(alpha, '\u{241b}', one_of(['\"', 'n'])).parse_next(i)
         }
         assert_eq!(esc3("ab␛ncd;"), Ok((";", "ab␛ncd")));
     }
@@ -676,7 +680,7 @@ mod complete {
     fn test_escaped_error() {
         fn esc(s: &str) -> IResult<&str, &str> {
             use crate::ascii::digit1;
-            escaped(digit1, '\\', one_of("\"n\\")).parse_next(s)
+            escaped(digit1, '\\', one_of(['\"', 'n', '\\'])).parse_next(s)
         }
 
         assert_eq!(esc("abcd"), Ok(("abcd", "")));
@@ -1377,7 +1381,7 @@ mod partial {
 
     fn digit_to_i16(input: Partial<&str>) -> IResult<Partial<&str>, i16> {
         let i = input;
-        let (i, opt_sign) = opt(one_of("+-")).parse_next(i)?;
+        let (i, opt_sign) = opt(one_of(['+', '-'])).parse_next(i)?;
         let sign = match opt_sign {
             Some('+') | None => true,
             Some('-') => false,