`&str` is convenient but
- It reads oddly, lookling like they are sequenced
- No good way to warn people about the performance without just removing
  it

Re-measured to performance to be sure:
```
contains_token/str/canada
                        time:   [15.830 ms 15.851 ms 15.874 ms]
                        thrpt:  [135.24 MiB/s 135.44 MiB/s 135.61 MiB/s]
contains_token/slice/canada
                        time:   [2.1956 ms 2.2033 ms 2.2111 ms]
                        thrpt:  [970.92 MiB/s 974.35 MiB/s 977.76 MiB/s]
```
All other impls are in the ballpark of `slice`

Fixes #226